requirements.txt
django
djangorestframework
django-redis
PyJWT
rsa
bcrypt
pytimeparse
mysqlclient
django-cors-headers